requirements.txt
cloudpathlib[s3]==0.16.0
pandas==2.1.4
scikit-learn==1.3.2
# https://docs.neptune.ai/integrations/sklearn/
neptune-sklearn==2.1.0
atenea-lib-neptune>=1.2.1
awswrangler==3.4.2
requests-aws4auth==1.2.3
jsonpath-ng==1.6.0
